refactor: rename module and update references from clients to client
