+++ anthropic-ai-sdk/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod clients;
+pub mod client;
 pub mod message_batches;
 pub mod messages;
 pub mod models;
