+++ examples/messages/count-message-tokens/src/main.rs
@@ -1,4 +1,4 @@
-use anthropic_ai_sdk::clients::AnthropicClient;
+use anthropic_ai_sdk::client::AnthropicClient;
 use anthropic_ai_sdk::types::message::{
     CountMessageTokensParams, Message, MessageClient, MessageError, Role,
 };
